refactor(peers): lift the peer staging + contract stores out of the api gate
Phase 3 prep for peer-agent goal in `octos chat`, mirroring the Phase 0
autonomy extraction. No logic changed — only module placement and item
visibility. `api::ui_protocol` reaches everything back through a re-export
(`ui_protocol_approvals`, ...) and a `use crate::peers::*` glob, so every
existing serve/WS call site resolves unchanged.

Why: the whole master<->peer loop hinges on ONE process-global registry.
`contract_stores()` is a `OnceLock` holding the pending-approval /
pending-question maps; a peer's own requester registers its oneshot there and
the master's `peer_respond` resolves it from the same map. `octos serve` got
that for free because both halves lived behind the `api` feature. `octos chat`
hosts its peers in the SAME process and needs the identical guarantee without
axum, `AppState`, or `WsConnection` — none of which these stores ever touched.

  crate::contracts  (was api/ui_protocol_{approvals,questions,scope,diff,sanitize}.rs)
      UiProtocolContractStores + contract_stores()

  crate::peers      (was ~2.4k lines of api/ui_protocol.rs)
      addressing (peer_slug_is_safe / name_to_slug / staged_peer_dir /
      resolve_peer_name_to_slug), fd-anchored peer_io, stage_peer +
      build_peer_handoff_callback, the blackboard reader + peer_list renderer,
      the parked-prompt projection + peer_respond_resolve, and the
      process-global peer wire registry.

Everything AppState/WebSocket-shaped stayed put: register/evict_peer_wire_session,
wake_master_on_peer_awaiting_input, write_peer_result_if_peer_session, the
peer/prepare + peer/gather RPC handlers, fleet synthesis, the SessionRuntime
model-lane builders, and build_peer_close_callback.

Verified: no test lost — the exact multiset of `#[test]`/`#[tokio::test]`
function names is identical to the base commit (2790 == 2790).
